# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Golf Court Reservation API",
        "description": "Documentación de la API para el sistema de reservas de canchas de golf.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Obtener listado de jugadores",
                "description": "Retorna una lista paginada de jugadores con filtros opcionales por nombre, apellido y club_id.",
                "operationId": "getPlayersList",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página (máximo 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre del jugador",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Filtrar por apellido del jugador",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "club_id",
                        "in": "query",
                        "description": "Filtrar por ID del club asociado",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Crear un nuevo jugador",
                "description": "Registra un nuevo jugador en el sistema asociado a un club.",
                "operationId": "storePlayer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "club_id",
                                    "name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "club_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Alejandra"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Zepeda"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alejandra@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+34600123456"
                                    },
                                    "handicap": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12.5
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jugador creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/players/{player}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Mostrar un jugador específico",
                "description": "Retorna la información detallada de un jugador mediante su ID.",
                "operationId": "getPlayerById",
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Actualizar un jugador existente",
                "description": "Modifica los datos de un jugador existente.",
                "operationId": "updatePlayer",
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "club_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Alejandra Actualizada"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Zepeda"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+346667788"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alejandra.nueva@example.com"
                                    },
                                    "handicap": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jugador actualizado con éxito"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Eliminar un jugador",
                "description": "Elimina un jugador del sistema.",
                "operationId": "deletePlayer",
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Jugador eliminado exitosamente (sin contenido)"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
...
    },
...
}